move codemirror editor components
